Safari-611.1.10/LayoutTests/fast/dom/Element/mozilla-dom-base-tests/test_bug364092.xhtml

<?xml version="1.0"?>
<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=364092
-->
<head>
    <meta charset="utf-8" />
    <script src="../../../../resources/js-test-pre.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=364092">Mozilla Bug 364092</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<div id="test1" foo="foo"/>
<pre id="test">
<script class="testbody" type="text/javascript">
/** Test for Bug 364092 **/

  var test1 = document.getElementById("test1");
  var attrNode = test1.getAttributeNode("foo");
  function mutationHandler(aEvent) {
    mutationEvent = aEvent;
    shouldBe('attrNode', 'mutationEvent.relatedNode');
    shouldBeFalse('test1.hasAttribute("foo")');
  }

  function runTest() {
    test1.addEventListener("DOMAttrModified", mutationHandler, true);
    test1.removeAttributeNode(attrNode);
    test1.removeEventListener("DOMAttrModified", mutationHandler, true);
  }

   runTest();
</script>
</pre>
<script src="../../../../resources/js-test-post.js"></script>
</body>
</html>
